requirements.txt
asgiref==3.7.2
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
cryptography==42.0.2
defusedxml==0.7.1
Django==4.2.10
django-allauth==0.61.1
django-appconf==1.0.6
django-classy-tags==4.1.0
django-crispy-forms==1.14.0
django-easy-maps==1.1.3
geographiclib==2.0
geopy==2.4.1
idna==3.6
oauthlib==3.2.2
pillow==10.2.0
pycparser==2.21
PyJWT==2.8.0
python-dotenv==1.0.0
python3-openid==3.2.0
razorpay==1.4.1
requests==2.31.0
requests-oauthlib==1.3.1
sqlparse==0.4.4
typing_extensions==4.9.0
urllib3==2.2.0
